moved socket context to context folder
