+++ src/component/context/ProfileContext.jsx
@@ -1,5 +1,5 @@
 import React, { createContext, useState, useEffect, useMemo, useCallback } from 'react';
-import { useSocket } from '../../hooks/useSocket.jsx';
+import { useSocket } from './useSocket.jsx';
 import { toast } from 'sonner';
 import axios from "../../utils/axiosInstance.js";
