+++ src/pages/Signup.jsx
@@ -6,7 +6,7 @@ import signUpValidator from '../validators/signUpValidator.js';
 import { HiOutlineMail, HiLockClosed, HiOutlineUserCircle } from "react-icons/hi";
 import { toast } from 'sonner';
 import UserIcon from '../component/icons/UserIcon.jsx';
-import { useSocket } from '../hooks/useSocket.jsx';
+import { useSocket } from '../component/context/useSocket.jsx';
 
 const initialFormData = { firstName: "", lastName: "", email: "", password: "", confirmPassword: "" };
 const initialFormError = { firstName: "", lastName: "", email: "", password: "", confirmPassword: "" };
